% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_olleyes_visualFields_pdfs.R
\name{combine_visual_field_pdfs}
\alias{combine_visual_field_pdfs}
\title{Combine Visual Field PDFs}
\usage{
combine_visual_field_pdfs(
  file_source,
  olleyes_file_csv,
  number_from_OE,
  file_names
)
}
\arguments{
\item{file_source}{The directory where the CSV file is located.}

\item{number_from_OE}{An integer to index the visual field data.}

\item{file_names}{A vector of filenames for the generated PDFs.}
}
\value{
The path to the combined PDF file.
}
\description{
This function reads visual field data from a CSV file, generates individual PDFs,
converts them to images, combines them side by side, and saves them as a new PDF file
with multiple pages.
}
